+++ test/bundler.ts
@@ -1,18 +1,19 @@
 import { promises as fs } from 'fs'
-import { resolve } from 'path'
-import { fileURLToPath } from 'url'
+import { join, resolve } from 'path'
+import { fileURLToPath, pathToFileURL } from 'url'
 
 import test from 'ava'
 import tmp from 'tmp-promise'
 
 import { BundleError } from '../src/bundle_error.js'
-import { bundle } from '../src/bundler.js'
+import { bundle, BundleOptions } from '../src/bundler.js'
 
 const url = new URL(import.meta.url)
 const dirname = fileURLToPath(url)
+const fixturesDir = resolve(dirname, '..', 'fixtures')
 
 test('Produces a JavaScript bundle and a manifest file', async (t) => {
-  const sourceDirectory = resolve(dirname, '..', 'fixtures', 'project_1', 'functions')
+  const sourceDirectory = resolve(fixturesDir, 'project_1', 'functions')
   const tmpDir = await tmp.dir()
   const declarations = [
     {
@@ -20,7 +21,16 @@ test('Produces a JavaScript bundle and a manifest file', async (t) => {
       path: '/func1',
     },
   ]
-  const result = await bundle([sourceDirectory], tmpDir.path, declarations)
+  const result = await bundle([sourceDirectory], tmpDir.path, declarations, {
+    basePath: fixturesDir,
+    importMaps: [
+      {
+        imports: {
+          'alias:helper': pathToFileURL(join(fixturesDir, 'helper.ts')).toString(),
+        },
+      },
+    ],
+  })
   const generatedFiles = await fs.readdir(tmpDir.path)
 
   t.is(result.functions.length, 1)
@@ -40,7 +50,7 @@ test('Produces a JavaScript bundle and a manifest file', async (t) => {
 })
 
 test('Produces only a ESZIP bundle when the `edge_functions_produce_eszip` feature flag is set', async (t) => {
-  const sourceDirectory = resolve(dirname, '..', 'fixtures', 'project_1', 'functions')
+  const sourceDirectory = resolve(fixturesDir, 'project_1', 'functions')
   const tmpDir = await tmp.dir()
   const declarations = [
     {
@@ -49,9 +59,17 @@ test('Produces only a ESZIP bundle when the `edge_functions_produce_eszip` featu
     },
   ]
   const result = await bundle([sourceDirectory], tmpDir.path, declarations, {
+    basePath: fixturesDir,
     featureFlags: {
       edge_functions_produce_eszip: true,
     },
+    importMaps: [
+      {
+        imports: {
+          'alias:helper': pathToFileURL(join(fixturesDir, 'helper.ts')).toString(),
+        },
+      },
+    ],
   })
   const generatedFiles = await fs.readdir(tmpDir.path)
 
@@ -71,7 +89,7 @@ test('Produces only a ESZIP bundle when the `edge_functions_produce_eszip` featu
 })
 
 test('Adds a custom error property to user errors during bundling', async (t) => {
-  const sourceDirectory = resolve(dirname, '..', 'fixtures', 'invalid_functions', 'functions')
+  const sourceDirectory = resolve(fixturesDir, 'invalid_functions', 'functions')
   const tmpDir = await tmp.dir()
   const declarations = [
     {
@@ -110,4 +128,55 @@ test('Does not add a custom error property to system errors during bundling', as
   } catch (error: unknown) {
     t.false(error instanceof BundleError)
   }
+})
+
+test('Uses the cache directory as the `DENO_DIR` value if the `edge_functions_cache_deno_dir` feature flag is set', async (t) => {
+  const sourceDirectory = resolve(fixturesDir, 'project_1', 'functions')
+  const outDir = await tmp.dir()
+  const cacheDir = await tmp.dir()
+  const declarations = [
+    {
+      function: 'func1',
+      path: '/func1',
+    },
+  ]
+  const options: BundleOptions = {
+    basePath: fixturesDir,
+    cacheDirectory: cacheDir.path,
+    importMaps: [
+      {
+        imports: {
+          'alias:helper': pathToFileURL(join(fixturesDir, 'helper.ts')).toString(),
+        },
+      },
+    ],
+  }
+
+  // Run #1, feature flag off: The directory should not be populated.
+  const result1 = await bundle([sourceDirectory], outDir.path, declarations, options)
+  const outFiles1 = await fs.readdir(outDir.path)
+
+  t.is(result1.functions.length, 1)
+  t.is(outFiles1.length, 2)
+
+  await t.throwsAsync(() => fs.readdir(join(cacheDir.path, 'deno_dir')))
+
+  // Run #2, feature flag on: The directory should be populated.
+  const result2 = await bundle([sourceDirectory], outDir.path, declarations, {
+    ...options,
+    featureFlags: {
+      edge_functions_cache_deno_dir: true,
+    },
+  })
+  const outFiles2 = await fs.readdir(outDir.path)
+
+  t.is(result2.functions.length, 1)
+  t.is(outFiles2.length, 2)
+
+  const denoDir2 = await fs.readdir(join(cacheDir.path, 'deno_dir'))
+
+  t.true(denoDir2.includes('deps'))
+  t.true(denoDir2.includes('gen'))
+
+  await fs.rmdir(outDir.path, { recursive: true })
 })