+++ test/fixtures/project_1/functions/func1.ts
@@ -1,1 +1,9 @@
-export default async () => new Response('Hello')
+import { greet } from 'alias:helper'
+
+import { echo } from '../../helper.ts'
+
+export default async () => {
+  const greeting = greet(echo('Jane Doe'))
+
+  return new Response(greeting)
+}